Shrink interface
* Removal of UDP protocol
* Removal of async interface
* Removal of the reader interface
* No more crashes on timeouts
  * Connection errors are returned as values
  * Allows clients to implement much finer control over what they need
   to do for back-pressure.

The UDP interface, whilst useful for a few select situations, has a
propensity for losing metrics, confusion and pain. It is gone. People
needing UDP can pin before this and have their tyranny if they wish.

The async interface was a weird choice and probably not very useful for
a bunch of people. There's a socket being written to in an
unbounded (though randomly dropped) way. It seems this could lead to
problems. It's trivial to add a small wrapper which makes it async
anyway, there's no need for the library to provide this. It's gone.

The reader/query interface, whilst obviously a cool feature, is of
dubious value outside of analytics and dashboards. Meh. It's gone.